requirements.txt
python-jobspy==1.1.59
requests==2.32.3
supabase==2.6.0
